Centralize verbosity level management
Added a new nsiqcppstyle_outputer class.
This class is instantiated as a global variable.
It is in charge of persisting the verbosity level.
It also exposes a Print method which will only print a given text
if it's defined verbosity level matches the one set by the user.

The following levels are defined:
   - Verbose (only displayed if the verbose option as been enabled)
   - Default (skipped if the "--ci" option is passed)
   - Ci (currently always displayed)
   - Error (always displayed)