Fixup multiple crashes (#57)
* Fixup crash if no --var param is passed

If no var is passed, the args.var will be set to None.
Trying to iterate on None throws the following exception:
    TypeError: 'NoneType' object is not iterable

The function now only processes the list if it's defined

* Fixup crash when filefilter contains empty lines

The RemoveOuterQuotes function did not check the line length,
resulting in the following error if an empty line is present:
    string index out of range

The function now only checks string that are at least 2 char long

* Fixup crash when filefilter contains comments

The _ProcessFilterLine function returned None when a comment was found
instead of the filter.
The caller replaces the filter with the function return value.
The next call then crashed with the following error:
    'NoneType' object has no attribute 'AddCppChecker'

The filter is now returned in this case as well

* Add an application test

This new test runs the app as the user would with a bit of configuration

* Review: invert GetCliKeyValueMap condition

* Review: rework test resources